requirements.txt
fastapi
openai
litellm
python-dotenv
